feat: implement confirm modal

+++ src/components/Board/Board.tsx
@@ -4,15 +4,20 @@ import deleteBoard from '../../api/deleteBoard';
 import { AppContext } from '../../App';
 import { SET_BOARDS } from '../../data/constants';
 import { BoardsResponse } from '../../data/interfaces';
+import useConfirm from '../../utils/useConfirm';
 import './Board.scss';
 
 function Board({ id, title }: BoardsResponse) {
   const { boards, dispatchBoards } = useContext(AppContext);
+  const { isConfirmed } = useConfirm();
 
-  const handleDeleteBoard = () => {
-    const filtredBoards = boards.filter((board) => board.id !== id);
-    dispatchBoards({ type: SET_BOARDS, payload: filtredBoards });
-    deleteBoard(id);
+  const handleDeleteBoard = async () => {
+    const confirmed = await isConfirmed(`You sure about that?`);
+    if (confirmed) {
+      const filtredBoards = boards.filter((board) => board.id !== id);
+      dispatchBoards({ type: SET_BOARDS, payload: filtredBoards });
+      deleteBoard(id);
+    }
   };
 
   return (
